src/backend/requirements.txt
Flask==2.1.2
flask-cors==3.0.10
openai==0.21.0
python-dotenv==0.20.0
